# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BitaSoft Api Documentation",
        "description": "BitaSoft Api Documentation",
        "contact": {
            "name": "Bita Soft",
            "email": "bitasoft22@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/customer/details": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get list of Customers",
                "description": "Get list of Customers",
                "operationId": "fc3d0996fa5b0c88a8b216418a9a8120",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Store Customers in DB",
                "description": "Store Customers in DB",
                "operationId": "7300f3b633de085109c46a3606e6e1bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "country"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "country": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/details/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Customers Detail",
                "description": "Get Customers Detail",
                "operationId": "c77f65e6f08ba53fd1ce30e14cfdab54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Customers",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update Customers in DB",
                "description": "Update Customers in DB",
                "operationId": "f8e993194ce7bfcfdc17899d3a07c4b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Customers",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "country"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "country": {
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Delete Customers",
                "description": "Delete Customers",
                "operationId": "e398bff036caded653b8279dffe310e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Customers",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}